node_ptr is introduced
The crate is converted into no_std.

Although the core structure of defining a self referential collection via the expressive `Variant` definition is kept, a major refactoring is applied.

Lifetimes are removed from the collections which simplifies the api significantly. The safety is provided internally by `NodePtr` and `NodeIdx`.

Memory reorganization is left to the implementors through the `MemoryReclaim` trait.

Example implementations of singly and doubly linked lists are provided.